Added LightingStrike ability

+++ Scripts/Resources/UnitStatList.cs
@@ -10,7 +10,7 @@ public partial class UnitStatList : Resource
 {
     [Export] public Godot.Collections.Array<UnitStat> StatList = new Godot.Collections.Array<UnitStat>();
 
-    public bool TryGetStatValue(StatType statType, out int value)
+    public bool TryGetStatValue(StatType statType, out float value)
     {
         for (int i = 0; i < StatList.Count; i++)
         {
